поправил генерацию комментариев в фильмах при взаимодействии с кнопками сортировки (и в целом перерендере компонента попапа)
