+++ src/view/popup-view.js
@@ -1,24 +1,6 @@
 import {humanizeDuration, humanizeReleaseDate} from '../utils/utils';
-import {generateComment} from '../mock/comment';
 import AbstractView from './abstract-view';
 
-const createCommentTemplate = (comments) => comments.map((item) => {
-  const { emote, comment, author, date } = item;
-
-  return `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${emote}.png" width="55" height="55" alt="emoji-${emote}">
-    </span>
-    <div>
-      <p class="film-details__comment-text">${comment}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${author}</span>
-        <span class="film-details__comment-day">${date}</span>
-        <button class="film-details__comment-delete">Delete</button>
-      </p>
-    </div>
-  </li>`;
-}).join('');
 
 const createMoviePopupTemplate = (movie) => {
   const {
@@ -36,19 +18,18 @@ const createMoviePopupTemplate = (movie) => {
       duration,
       genres,
       description,
+      comments,
+      commentsListTemplate
     },
     userData: {
       isInWatchlist,
       isAlreadyWatched,
       isFavorite,
     },
-    comments
   } = movie;
 
   const formattedDuration = humanizeDuration(duration);
   const formattedReleaseDate = humanizeReleaseDate(releaseDate);
-  const commentsArray = comments.map(generateComment);
-  const commentsListTemplate = createCommentTemplate(commentsArray);
 
   const watchlistClassName = isInWatchlist
     ? 'film-details__control-button--watchlist film-details__control-button--active'
@@ -134,7 +115,7 @@ const createMoviePopupTemplate = (movie) => {
 
     <div class="film-details__bottom-container">
       <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsArray.length}</span></h3>
+        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
 
         <ul class="film-details__comments-list">
           ${commentsListTemplate}
